Jetty 8 package renamed.
