+++ h2o-jetty-8/src/main/java/water/webserver/jetty8/Jetty8DelegatingAuthenticator.java
@@ -1,4 +1,4 @@
-package water.webserver.jetty9;
+package water.webserver.jetty8;
 
 import org.eclipse.jetty.security.Authenticator;
 import org.eclipse.jetty.security.ServerAuthException;
