+++ h2o-jetty-8/src/main/java/water/webserver/jetty8/Jetty8Facade.java
@@ -1,4 +1,4 @@
-package water.webserver.jetty9;
+package water.webserver.jetty8;
 
 import water.webserver.iface.Credentials;
 import water.webserver.iface.H2OHttpView;
