+++ h2o-jetty-8/src/main/resources/META-INF/services/water.webserver.iface.HttpServerFacade
@@ -1,1 +1,1 @@
-water.webserver.jetty9.Jetty8Facade
+water.webserver.jetty8.Jetty8Facade
